--- a/optimum/exporters/onnx/convert.py
+++ b/optimum/exporters/onnx/convert.py
@@ -1209,11 +1209,9 @@ def onnx_export_from_model(
 
         from onnxslim import slim
 
-        onnx_models = glob.glob(os.path.join(output, "**", "*.onnx"), recursive=True)
-
-        for model in onnx_models:
-            slimmed_model = slim(model)
-            check_and_save_model(slimmed_model, model)
+        for onnx_file_path in onnx_files_subpaths:
+            slimmed_model = slim(onnx_file_path)
+            check_and_save_model(slimmed_model, onnx_file_path)
 
     # Optionally post process the obtained ONNX file(s), for example to merge the decoder / decoder with past if any
     # TODO: treating diffusion separately is quite ugly