--- a/optimum/utils/__init__.py
+++ b/optimum/utils/__init__.py
@@ -51,6 +51,7 @@ from .import_utils import (
     is_transformers_available,
     is_transformers_version,
     require_numpy_strictly_lower,
+    is_onnxslim_available,
 )
 from .input_generators import (
     DEFAULT_DUMMY_SHAPES,