--- a/setup.py
+++ b/setup.py
@@ -105,7 +105,7 @@ EXTRAS_REQUIRE = {
     "amd": "optimum-amd",
     "quanto": ["optimum-quanto>=0.2.4"],
     "dev": TESTS_REQUIRE + QUALITY_REQUIRE,
-    "tests": TESTS_REQUIRE + ["onnxslim"],
+    "tests": TESTS_REQUIRE + ["onnxslim>=0.1.53"],
     "quality": QUALITY_REQUIRE,
     "benchmark": BENCHMARK_REQUIRE,
     "doc-build": ["accelerate"],