--- a/tests/exporters/onnx/test_export_cli.py
+++ b/tests/exporters/onnx/test_export_cli.py
@@ -751,5 +751,4 @@ class OnnxCLIExportTestCase(unittest.TestCase):
             slim=True,
             device="cpu",
             variant=variant,
-            model_kwargs=model_kwargs,
         )