# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->
This pr introduces [onnxslim](https://github.com/inisis/OnnxSlim), a new high performance onnx optimization tool.

for gpt2 model, we can reduce number of operators and model size without losing accuracy, and applies to many other models.
```
optimum-cli export onnx --model gpt2 gpt2_onnx/ --slim
```

```
+------------------------------+------------------------------------------+------------------------------------------+
|             Add              |                   114                    |                   114                    |
|             Cast             |                    34                    |                    3                     |
|            Concat            |                   177                    |                   113                    |
|           Constant           |                   1054                   |                    0                     |
|       ConstantOfShape        |                    4                     |                    2                     |
|             Div              |                    37                    |                    25                    |
|            Equal             |                    2                     |                    2                     |
|            Expand            |                    2                     |                    2                     |
|            Gather            |                   186                    |                   185                    |
|             Gemm             |                    48                    |                    48                    |
|             Less             |                    1                     |                    1                     |
|            MatMul            |                    25                    |                    25                    |
|             Mul              |                    99                    |                    97                    |
|             Pow              |                    37                    |                    37                    |
|            Range             |                    1                     |                    1                     |
|          ReduceMean          |                    50                    |                    50                    |
|           Reshape            |                   150                    |                   145                    |
|            Shape             |                   250                    |                   100                    |
|            Slice             |                    76                    |                    14                    |
|           Softmax            |                    12                    |                    12                    |
|            Split             |                    12                    |                    12                    |
|             Sqrt             |                    61                    |                    25                    |
|           Squeeze            |                    52                    |                    2                     |
|             Sub              |                    27                    |                    27                    |
|             Tanh             |                    12                    |                    12                    |
|          Transpose           |                    61                    |                    61                    |
|          Unsqueeze           |                   271                    |                   187                    |
|            Where             |                    5                     |                    5                     |
+------------------------------+------------------------------------------+------------------------------------------+
|          Model Size          |                475.12 MB                 |                475.02 MB                 |
+------------------------------+------------------------------------------+------------------------------------------+
|         Elapsed Time         |                                        9.57 s                                       |
+------------------------------+------------------------------------------+------------------------------------------+
```


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?

## Who can review?
@IlyasMoutawwakil @xenova 
<!--
For faster review, we strongly recommend you to ping the following people:
- ONNX / ONNX Runtime : @fxmarty, @echarlaix, @JingyaHuang, @michaelbenayoun
- ONNX Runtime Training: @JingyaHuang
- BetterTransformer: @fxmarty
- GPTQ, quantization: @fxmarty, @SunMarc
- TFLite export: @michaelbenayoun
-->