@IlyasMoutawwakil 2025-05-12
@inisis the table is very cool ! is it possible to get the "diff" i.e. how many operators were removed per operator type and overall reduction in number of ops ?