@inisis 2025-05-12
> @inisis the table is very cool ! is it possible to get the "diff" i.e. how many operators were removed per operator type and overall reduction in number of ops ?

@IlyasMoutawwakil Currently, we don't support that, but this feature can be added, and we provides api  for doing so, and there will be different colors in terminal, green means less.

![image](https://github.com/user-attachments/assets/4f2fdcf1-7091-4eb0-8088-01cea967cf3b)