@IlyasMoutawwakil 2025-05-13
> can I use your CI server to do it

The CI runs on basic Github runners.