@xenova 2025-05-13
Just want to add my voice in support of onnxslim: it's an amazing library I use for all my Transformers.js models 🤗. So, having this built-in to Optimum could be really helpful too! 👍 

Happy to provide a review when ready too.